nginx: Bundle a copy of the default nginx.conf
We need this in case we want to completely overwrite /etc/nginx, which
is a good idea. (I think.)

The impetus here being that we don't really want to trust or force
ourselves to use the default configurations. This gives us even finer
control and lets us eventually slim things down a lot.